Reworded section about project budget.
This was motivated by reading this paragraph from the EdgeCase website: "EdgeCase offers development in units of two-week, two-person iterations. Our relationship lasts as long as you love our product." I'm not sure I've quite captured what I intended, but hopefully it's a bit of an improvement. I'm slightly concerned the "coming back for more" phrase is a bit cheesy!