The `BodyClass` DynaSnip was causing an error resulting in horrible HTML output.
The error was an uninitialized constant `Vanilla::Renderers::Ruby::DynaSnip`.

I noticed that the `Kind` and `LinkTo` DynaSnips were requiring 'vanilla/dynasnip', so I added the same require statement to `BodyClass`. This fixed the problem.

I took the liberty of adding the same require statement to the other DynaSnips missing it.