Limit the number of entries in the blog feed.
Our feed wasn't updated in FeedBurner; I tried to validated it using
the URL, but the validator timed out. Copying the contents of the
feed into the validator worked (and the feed is valid), so my theory
is that FeedBurner is timing out requesting the feed.

I did some profiling and found that rendering the feed is just
generally expensive because of the general render process. A quick
fix is to limit the number of entries in the feed.