Remove sitemap root snip workaround
This removes the workaround[1] for the sitemap and root snip, which
was fixed in vanilla-rb 1.17.3.

It's worth noting that having the name attribute of a snip differ
from the name of the file can lead to unexpected behaviour; for
this reason I've changed the name of the snip back to the default
("start"), and used the 'page_title' functionality to keep the
important behaviour (that the page `<title>` isn't 'start').

[1]: abbabf4859342d24e3a21dc4a554f32c4b305346