Requests for mocha/recap docs should be served from the file system.
I'm not sure this is the best solution, especially since we've still
got an AliasMatch directive in place along with a bunch of redirects
in the Apache configuration, but it does look like this means we get
404s instead of 500 errors when requesting non-existent files in either
of these docs directories. I think the 500s were happening because when
Apache couldn't find the file it asked the Vanilla app to serve them
which is what raised the 500.