Fix one of the broken images on build-caffeine-monitor page.
The link to the image was broken and the link to the "full size" image was
also broken. I've found a working version of the former, but there is no
equivalent for the latter. And in any case Flickr community guidelines
[1] now say that you should provide a link back the Flickr page for the
photo which is what I've done.

Note that all the other Flickr photos are broken on this page, but the
photos no longer appear to be public on Jase's Flickr account.

[1] http://www.flickr.com/guidelines.gne