Respond with 404 not 500 when a Dynasnip is accessed directly
It's possible to generate 500 errors by hitting e.g.
http://gofreerange.com/member. This seems undesirable, not least
because it generates an exception notification email.

The exception we were seeing was:

    ArgumentError: wrong number of arguments (0 for 1)

    vendor/gems/ruby/1.8/bundler/gems/vanilla-rb-586acf13dd8f/lib/vanilla/renderers/ruby.rb:35:in `handle'
    vendor/gems/ruby/1.8/bundler/gems/vanilla-rb-586acf13dd8f/lib/vanilla/renderers/ruby.rb:35:in `send'
    vendor/gems/ruby/1.8/bundler/gems/vanilla-rb-586acf13dd8f/lib/vanilla/renderers/ruby.rb:35:in `process_text'
    vendor/gems/ruby/1.8/bundler/gems/vanilla-rb-586acf13dd8f/lib/vanilla/renderers/base.rb:99:in `render_without_including_snips'
    vendor/gems/ruby/1.8/bundler/gems/vanilla-rb-586acf13dd8f/lib/vanilla/renderers/base.rb:87:in `render'
    vendor/gems/ruby/1.8/bundler/gems/vanilla-rb-586acf13dd8f/lib/vanilla/app.rb:70:in `render'

A better fix might be to make these snips not directly routable,
but I think that would involve changes to Vanilla, so I've avoided
that for now. Also I'm not sure if some of them *need* to be directly