Fix Encoding::CompatibilityError.
We were seeing exceptions on `/linda-reads-your-numbers`:

    Encoding::CompatibilityError: incompatible encoding regexp match (Windows-31J regexp with UTF-8 string)

The problem was with the accented character in `cafés` in combination with
BlueCloth. It seems as if we're using an old version of BlueCloth via
vanilla.rb. Rather than considering upgrading BlueCloth, it seems more
sensible to standardize on Kramdown which doesn't suffer from this encoding
problem and which we're already using in most places.