Remove redundant printer-kit snip.
We've been redirecting [1] the URL for this page for some time, so it's surplus
to requirements.

Although there are still references to the `/printer-kit` path on pages on our
site, these will be redirected to `exciting.io`, so I don't think we need to do
anything about them urgently. At some point it might be nice to add an "Update"
paragraph at the top of relevant blog posts to point out that things have
changed, but I think that can wait for the moment.

[1] https://github.com/freerange/site/blob/master/config/apache/gofreerange.com.conf#L86