Include year in blog post datestamp.
I was talking to @tomstuart about whether I ought to add updates to old blog
posts which have been rendered out-of-date by events. He suggested that
although adding an update would be nice, it's more important to make it clear
when the blog post was written so that it can be read in that context.

I noticed that our blog posts did not include the year in their datestamp and
so I decided to remedy that here.