Remove redundant printer-questions snip.
We've been redirecting [1] the URL for this page for some time, so it's surplus
to requirements.

This snip was originally added in [2] along with the images which I'm also
removing in this commit.

It doesn't look as if there are any references to the `/printer-questions` path
in pages on our website, but if there are, they'll be redirected to
`exciting.io` anyway.

[1] https://github.com/freerange/site/blob/master/config/apache/gofreerange.com.conf#L85
[2] d5dbdb1fb673eda36043067445496119bb8e62c9