Remove redundant hello-printer snip.
We've been redirecting [1] the URL for this page for some time, so it's surplus
to requirements.

I've also removed the images which were associated with this snip when it was
added in [2].

Although there are still references to the `/hello-printer` path on pages in our
website, these will be redirected to `exciting.io`, so I don't think we need to
do anything about them urgently. At some point it might be nice to add an
"Update" paragraph at the top of relevant blog posts to point out that things
have changed, but I think that can wait for the moment.

[1] https://github.com/freerange/site/blob/master/config/apache/gofreerange.com.conf#L84
[2] 157324f3e00955a4f8ff5777e5d80e2c23218059