Remove redundant printer project snip.
We've been redirecting [1] the URL for this page for some time, so it's surplus
to requirements.

I've also removed the image which was associated with this snip in [2].

Although there are still references to the `/printer` path on pages in our
website, these will be redirected to `exciting.io`, so I don't think we need to
do anything about them urgently. At some point it might be nice to add an
"Update" paragraph at the top of relevant blog posts to point out that things
have changed, but I think that can wait for the moment.

[1] https://github.com/freerange/site/blob/1f508583ab71be4a29c682e094478c4a1a1d6938/config/apache/gofreerange.com.conf#L87
[2] 179cb112580efab45d7cee2c8886183da7b79b1c