Allow non-blog pages to set page title.
Previously pages which used the standard layout had their page title
hard-coded. I've changed the standard layout to use the value of the
`page_title` attribute.

Previously the home page `page_title` attribute value was being ignored,
so in order to preserver the original page title for the home page I've
changed the value of the `page_title` attribute.

In case there are any other pages using the standard layout which do not
have a `page_title` attribute value set, I've set the default to be the
same as the original hard-coded value.

So hopefully this commit does not alter any behaviour, but it does give
us the possibility of setting the page title for new non-blog pages
which we create.