Add a basic page for the Show and Tell events.
Note that I had to use HTML for the bullet points, because I needed
to generate paragraph tags inside the list item tags to avoid the
bullet points looking tiny due to the CSS styles for non-blog pages
not being setup very well.

Even so the page doesn't look great, but hopefully it'll do for now.

Note that I set the `is_page` attribute so that this page is included
in the sitemap. Hopefully since my previous commit, the `page_title`
attribute value will also be honoured.