Use default page title if snip is missing.
Having added the page_title snip to the standard layout in [1], I started
seeing the following error:

    A NoMethodError occured: undefined method `page_title' for nil:NilClass

For URLs which should've been returning a 404 Not Found e.g.

    http://gofreerange.com/apple-touch-icon-precomposed.png

This was because the standard layout is used even for rendering 404 pages
when there is no snip available.

My change is to only call `Snip#page_title` if a snip is available.

This stuff is all very confusing. For example there are two `PageTitle`
dynasnips - one in `soups/dynasnips` and one in `soups/system`.

[1]: 3afe11fe026a69cad571e996f26192247b84cf01