Handle invalid parameter errors.
We were seeing exceptions like the one below in production which were resulting
in 500 Internal Server Error responses which the `Rack::MailExceptions`
correctly emails us about:

    A ArgumentError occured: invalid %-encoding (<%MYTEST)

    ===================================================================
    Request Body:
    ===================================================================

      <%MYTEST=REquEst("autoshell"):EvaL(MYTEST)%>

I was able to reproduce the problem using a command like this on my local
development environment:

    curl -v -XPUT --data "<%MYTEST=REquEst("autoshell"):EvaL(MYTEST)%>" "http://localhost:9292/"

Upgrading Rack to v1.6.0 means that this type of request now raises a more
specific exception: `Rack::Utils::InvalidParameterError`. Rails handles [1]
this more specific exception and responds with a "400 Bad Request" response.

I've attempted to reproduce similar behaviour using some custom Rack
middleware loosely based on the `rack-robustness` middleware gem [2]. I've
inserted this middleware just in front of the `Rack::MailExceptions`
middleware so that the app responds with a "400 Bad Request" response instead
of emailing us with about a "500 Internal Server Error".

@chrisroos: I'm not sure this is the best solution and it does feel as if we're
reinventing the wheel a bit. It makes me wonder again whether we'd be better
off with a Rails app! But this solution does mean that we won't continually be
bothered with emails about such exception emails.

[1]: https://github.com/rails/rails/commit/75eaefcc2f7c7526be389702c4802a73d4d934e9
[2]: https://github.com/blambeau/rack-robustness