Use Airbrake::Rack instead of Rack::MailExceptions.
This commit replaces the `Rack::MailExceptions` middleware with
`Airbrake::Rack` for the reporting of exceptions.

We're already using Airbrake to report exceptions from our `video` and
`webhooks` apps to our instance of Errbit running on Heroku. So it makes sense
to bring this app into line.

Errbit makes it easier to manage exceptions when they occur e.g. commenting on
them and marking them as resolved. Errbit notifies us by email of any
exceptions, so none of the original behaviour is lost.