Remove use of WidowFix
This was causing problems with the bullet point links on this page [1].

We don't really understand what it's doing and we haven't seen anywhere
that suffers unduly from not having it running.

Also the JS seems really buggy.

Removing it seems like the best option at the moment. We can always reintroduce
it if we see a problem in the future.

[1]: http://gofreerange.com/week-336