Change list-style-type of the ordered list on blog index page
At some point the numbers in the ordered list items might've matched up with
the numbers of the week notes, but they certainly don't any more.

Since the order of the blog posts is semantically meaningful, I've elected to
just change the `list-style-type` to `disc` instead of switching the ordered
list to an unordered list.