Allow list-of dynasnip to list multiple kinds of snip
I'm planning on converting the Show & Tell event pages to be of kind
"show-and-tell" so I can list them dynamically on the Show & Tell index page.
However, in order to do that and keep those event pages in the blog index
list, we'll need to be able to include multiple kinds of pages i.e. both
kind "blog" and kind "show-and-tell".