Upgrade vanilla & soup
Fixes #34.

I've added a constraint for `rack`, because newer versions require a newer
version of Ruby.

I've added a constraint for `vanilla`, because it was the only way I could get
bundler to use the new beta version.

I have confirmed that the bug is fixed in that the request returns a 404 vs 500:

    $ curl -I http://localhost:9292/javascripts/h.html
    HTTP/1.1 404 Not Found
    Content-Type: text/html
    Content-Length: 0
    Server: WEBrick/1.3.1 (Ruby/2.1.6/2015-04-13)
    Date: Fri, 15 Jul 2016 11:14:50 GMT
    Connection: Keep-Alive