Add updated_at to sometimes-success-means... blog post
The lack of updated_at metadata meant that Vanilla would use the mtime
timestamp of the file. This in turn meant that some recent non-content
changes to this file caused it to show up in the RSS feed.

I used the following script to determine that this was the only blog
post / week notes post / week links post or show and tell post that was
lacking the updated_at:

```
Dir['soups/**/*.snip*'].each do |snip_file|
  snip_content = File.read(snip_file)
  next if snip_content =~ /updated_at/

  p snip_file
end
```