Only display three significant projects on the home page
Trello card: https://trello.com/c/vJA6vOFn

I've extracted all the projects into a separate page and link to it
from the bottom of the list on the home page.

Note that I've chosen to include the projects on the home page in
the separate list, because this made more sense to me i.e. the
separate page is intended to be a full list of projects.