Add a show-and-tell-layout snip and use it in all S&T pages
The main thing that this gives us is a link (the GFR logo) back
to the S&T index page from every S&T page.

This layout is heavily based on the "blog-layout" snip and so it
introduces a bunch of duplication, but I couldn't work out a way
to avoid this.

The fact that I've added the layout setting to the "show-and-tell-nn"
snip should mean that we don't forget to add it to any new S&T pages.