Make BlogRenderer use template based on kind of snip
Also add a new "show-and-tell-template" snip heavily based on the
"blog-template" snip. This allows us to list links to S&T event pages
in the "Recent" sidebar when we are on an S&T page, but recent blog
articles when we are on a blog page.

Again this introduces rather more duplication than I would like and
it's less obvious that BlogRenderer is still a good name for that
class. However, without a wholesale restructuring I can't see a better
way to do this.