Add a short section about S&T events to the home page
Trello card: https://trello.com/c/Qx7qFqx7

The main reason behind this is to provide a link from the home page
to the S&T events index page. Ideally this would be more of a
navigation-type link, but I have no idea to add one of those to the
current layout, so this will have to do.

I chose not to add this new section as a "project" as suggested by
the Trello card, because I didn't feel as if it fitted in very well
there.