Avoid overloading kind attribute to mark snips as draft
I'm pretty sure the only effect of marking a snip as draft is
to not include it in the the lists of links to snips generated
by the `list-of` snip.

This commit introduces a new `draft` attribute to mark snips as
draft instead of overloading the `kind` attribute. Snips with
the `draft` attribute set to `true` will be excluded from lists
of links to snips generated by the `list-of` snip.

The benefit of this approach is that other code can *rely* on
the `kind` attribute always being meaningful even when the snip
is in the "draft" state.

I've updated the Rake tasks and their corresponding templates to
work with the new approach. I've also updated the README as
suggested by @chrisroos.

I think that snips in directories with a leading underscore
(e.g. "soups/blog/_drafts") are not available in the "soup",
but for consistency I have marked them as draft anyway.