Render show-and-tell pages
This is achieved by adding a soup for the relevant directory to
the composite soup. Also by making the filter for the navigation
side-bar specific to the kind of snip. The latter mirrors the
behaviour provided by the choice of blog-template.snip.haml vs
show-and-tell-template.snip.haml.

Note that snip tags are still not honoured within these pages.