Make values of render_as consistently capitalised across the site
These values of the `render_as` attribute are used to select the
custom `Vanilla::Renderers::Blog` renderer. Although it doesn't
seem to have caused a problem this subset of pages has the attribute
set to the non-capitalised value, "blog", rather than the capitalised
value, "Blog", used in the majority of pages and which matches the
renderer class name.