Hacky implementation of atom feed using ratom gem
The implementation is closely based on the `Feed` dyna-snip
and `Vanilla::AtomFeed`. However, I was unable to use that
directly, because it relies on an instance of the vanilla-rb
app to render the snip.

The most significant difference in my implementation is that
the snip is rendered *without* its layout. This is because it
seems odd to do this within the atom feed - no need to include
author photo, Disqus comments, or navigation side-bar. Also
doing so would've made the implementation more complex.

Another internal difference is that I haven't done any custom
formatting of timestamps (c.f. `Vanilla::AtomFeed#atom_time`).
The output seems to be the same for significantly less complex
code. It might be that this is just working because of Rails'
extensions to Date & Time.

In particular the entry ID generated by vanilla seemed needlessly
complicated - I've just converted the timestamp to a date which
seems to give the same output.

The one remaining issue with the output is that the internal links
inside the content element only include the path and not the
protocol or host, c.f. `Vanilla::AtomFeed#externalise_links`.
I'm not sure how important this is - it's possible that Feedburner
or feed readers take care of this by default.