Avoid using vanilla-rb helper for non-existent 'printer' snip
When a snip doesn't exist, the `l.snip.rb` dyna-snip currently
renders the link OK, but adds a CSS class ("missing"). However,
this CSS class doesn't appear to have any visual effect. Thus
there is scope for this to silently render a link to a
non-existent page.

However, in this case, the link is intercepted by Apache and
redirected to a URL on exciting.io. But it feels better to
link directly to this external URL and avoid the vanilla-rb
`l` helper method altogether.