Make url_to & l helper methods more like their vanilla-rb equivalents
I've realised I made a mistake in the commit with title:
"Incorporate blog-template markup into blog layout". The `UrlTo`
dyna-snip which I referenced was in the `pristine_app` bit of
the vanilla-rb gem and is not actually present in the GFR website.

Instead we want `PagesController#url_to` to be very similar to
`Vanilla::Routing#url_to`. The main differences are:

* Special handling for root snip
* CGI escape snip name when generating path
* Append snip part when generating path

As it turns out, I don't think any of these scenarios currently
arise on the GFR website, but I thought it was worth being thorough.

Also since we're now [1] failing fast in the `l` helper method in the
vanilla-rb version of the website if the snip doesn't exist, I think
we should do the same in `ApplicationHelper#l`.

[1]: d4e6203618abc89ed8e62153babd650ea63e65af