Remove some links immediately before apostrophes
These apostrophes are correctly converted into right single quotes
by vanilla-rb and the Kramdown renderer. In this case, I believe
the snip is first converted to HTML by `Vanilla::Renderers::Kramdown`
and then the snip tags are processed by vanilla-rb.

    Kramdown::Document.new(%{{l foo, I}'ve}).to_html
    => "<p>{l foo, I}’ve</p>\n" (rsquo)

In the `rails-spike` branch I'm processing ERB tags first and then
converting to HTML using Kramdown. In that case, for some reason
this combination results in a *left* single quote which looks wrong.
Given that the author of the two articles is clear, I don't think
there's much value in retaining the links and so I'm happy to remove
them.

    Kramdown::Document.new(%{<a href="/foo">I</a>'m}).to_html
    # => "<p><a href=\"/foo\">I</a>‘m</p>\n" (lsquo)

Interestingly if the first letter after the apostrophe is the
letter "s" then a right single quote is correctly generated.

    Kramdown::Document.new(%{<a href="/foo">Tom</a>'s}).to_html
    # => "<p><a href=\"/foo\">Tom</a>’s</p>\n" (rsquo)