Remove redundant ":render_as: markdown" metadata from markdown snips
The `.markdown` file extension is sufficient to indicate how these
snips should be rendered.

I believe these metadata entries are the remnants of a time when
the underlying Bluecloth renderer was used by default for rendering
markdown, but this could be overridden to use Kramdown by setting
the `render_as` attribute.