Add class to allow black-box regression testing
The Spider class is made up of two main components.

The server class runs the web app locally from the command line in a
separate process. The idea of running it from the command line is to be
as "black box" as possible so that it's easy to change this test if we
change the web app substantially, e.g. to be a Rails app.
The `healthcheck_url` is used to determine whether the server has
started successfully.

The client class spiders the site as presented by the locally running
server, following internal links recursively. Link URLs are not
considered internal if they have a non-gofreerange.com domain or if they
use the 'https' scheme. The HTML responses are captured in the supplied
`artefacts_path`. The client ignores paths matching patterns in the
`ignore_paths` array. The client is primed with the links scraped from
the sitemap URL. The client will retry requests which do not return a
success HTTP status code; this is because I've been seeing some weird
intermittent exceptions like this one:

    NoMethodError: undefined method `name' for nil:NilClass

Requesting the same URL again somehow makes the exception go away!