Update regression test artefacts
Apart from the two exceptions below, at this point the Rails app is
rendering exactly the same content as the Vanilla app.

* The feed entries differ because the Rails app version only contains
the underlying snip content, i.e. it's not wrapped in the content from
the `blog-template` or `show-and-tell-template`. This seems like more
sensible behaviour, because the date & author are included in feed entry
attributes anyway, and there seems no value in including Disqus comments
or recent articles in each feed entry.

* The `blog-template` & `show-and-tell-template` pages differ, because
the dynasnips are not expanded. However, these snips were never intended
to be rendered independently. The HTML in these snips is now included in
the `blog` & `show-and-tell` layout templates in the Rails app. I plan
to remove the two snips in a later commit.