Use 'gfr' snip vs 'fr' everywhere for consistency
The former renders "Go Free Range" as a link to the home page while the
latter renders "Free Range" as a link also to the home page. I think we
mostly talk about ourselves as "Go Free Range" or "GFR" these days and
our company name is "Go Free Range Ltd" and our website is
gofreerange.com, so this feels like the most sensible option.

The main thing is that this makes the website more consistent. I'm happy
to change it if others disagree.