Remove unnecessary spaces between link & punctuation
The Rails version of the app doesn't appear to suffer from a problem [1]
where Markdown links followed by punctuation were broken. So we can
remove these redundant & confusing spaces.

The regression test artefacts indicate that all of the affected links
are still working after this change.

Note that it's possible this problem was actually fixed by the move from
`rdiscount` to `kramdown` or by the upgrade of `kramdown`.

[1]: https://github.com/freerange/site/issues/7