Use ... vs &hellip; in markdown documents
Kramdown converts "..." to a horizontal ellipsis HTML entity.

The changes to the regression test artefacts are all because the ID of
the heading elements no longer includes "hellip". I've checked that no
internal links currently reference these IDs, but it's possible that
other sites link to them. However, since the link will still take the
user to the correct page, I'm happy to make this change.