Attempt to fix a crawl error from Google Webmaster Tools
The following URL was giving a 403 Forbidden response:

http://gofreerange.com/mocha/docs/Mocha/ParameterMatchers/

I believe this was because directory indexes were disabled for the
`/home/freerange/docs/mocha` directory.

As far as I can see none of the "Linked from" pages reported by GWT
actually contain the aforementioned URL.

I did contemplate setting up a redirect to the following URL:

http://gofreerange.com/mocha/docs/Mocha/ParameterMatchers.html

However, I can't see the harm in enabling directory indexes and that
seems like a simpler fix.

The changes in this commit were applied once before as this commit [1].
However, I didn't realise I needed to run `cap deploy` as well as
`cap apache:update_config` to have the changes take effect on the
server.

[1]: f5131d2aa4a694fc86498d0c77a7327394f964d7