Redirect legacy Mocha doc URL to most relevant existing doc
I came across this as a crawl error in Google Webmaster Tools.

The `Mocha::AutoVerify` class is long since extinct and the closest
equivalent is now `Mocha::API`.

This makes me wonder whether we'd be better making the URLs for the docs
version-specific, or even relying on hosted solutions e.g. this site [1]

[1]: http://www.rubydoc.info/github/freerange/mocha