Avoid ul nested directly inside ul - invalid HTML
This fixes the nested `ul` element problem in the navigation sidebar on
individual blog post and S&T event pages. The outer `ul` has been
replaced with a `div` element and some `left-padding` has been applied
to this new `div` to preserve the existing layout/alignment of the page.

Note that the changes in the regression test artefacts are a bit
confusing, because html-tidy was converting the outer `ul` into a `div`
automatically and adding the following attribute:

    style="margin-left: 2em"

It therefore looks as if this attribute is being removed.

I'm happy that the changes to the regression test artefacts is
consistent with the change I was trying to make.

Fixes #61.