Make week note headings match page title
The heading is currently used by `ApplicationHelper#list_of` to
generate the text for each link in the list, but the logic to do this
feels pretty brittle - it just uses the first line of the content and
assumes that's the heading.

I think it would be better to rely on the snip metadata and the
`page_title` seems a pretty sensible option. This commit brings the
headings for the week links into line with their page titles so that
we can use the latter and see no change in a later commit.

In each case I've chosen the longer or more interesting of the heading
vs page title.

Note that the regression test artefacts do change in this commit as
expected.