Remove draft status from week-426
I'm confident that this page is intended to be public, because
@chrisroos tweeted [1] a link to it on 17 Mar 2017.

Currently a draft snip is available if accessed directly, so the only
consequence was that this was *not* included in the blog index page, in
the Atom feed (when it was in the most recent 10 posts) and in the
sitemap. This commit means that it will now be included in all these
things as appropriate.

I'm planning to prevent direct access to draft snips in a future commit,
and so making this not draft will avoid it being accidentally hidden.

The changes to the regression test artefacts are as I expected.

[1]: https://twitter.com/freerange/status/842793129141067777