Change ul -> ol in navigation side-bar
The `reversed` attribute is not allowed on a `ul` element and means
these pages are invalid HTML.

After a conversation with @chrisroos we decided that even though the
list items within the `ul` are not displayed with a number, they *are*
in reverse chronological order and so semantically it seems to make
sense to convert the `ul` element into an `ol` element.

In order to avoid any change in appearance, I've had to change the CSS
selector to match the new `ol` element.

Closes #77.