Remove unused JavaScript files
The only JavaScript file which is currently included in any page is
`public/javascripts/app.js`. This appears to be made up of a jQuery
v1.4.3 (minified), the jQuery.ScrollTo plugin v1.4.2 (minified), and
some custom JavaScript (non-minified). This JavaScript is currently
providing two bits of behaviour:

* Adds backlinks to footnotes which scroll the page back up to the
footnote link taking account of the navigation bar at the top of
the page.

* Adds a CSS class to the `body` element representing the day of the
week. This code makes use of the jQuery.ChronoClass plugin [1]
(non-minified). There is a CSS rule which makes use of the CSS class to
display a background image on Thursdays (`/images/bg/thurs.jpg`).

Here are some notes on the files I'm removing in this commit:

* application.js - the behaviour in this file was disabled many moons
ago in this commit [2].

* hashgrid.js - I think this [3] was used as a design aid in development
by @jasoncale, but I'm pretty confident it's not used in production.

* jquery.scrollTo.js - as I mentioned above, this *is* being used, but
this minified version of the plugin has been incorporated into app.js.

* vanilla.js - this looks as if it's designed to allow editing of a
vanilla-based website, so this should've been removed when we switched
the website to Rails.

* jquery.autogrow-textarea.js - it looks as if this was only used by
vanilla.js.

* jquery.enumerable.js - this was introduced in this commit [4] and it
looks as if it was used by application.js.

* jquery-ui-1.8.2.custom.min.js - I'm not sure where this was used, but
it was introduced in the same commit [4] and was probably also used by
application.js.

[1]: https://github.com/freerange/jQuery-ChronoClass
[2]: 381bc0fe3585cd02e8a8b821795521cc0a4981c6
[3]: https://github.com/dotjay/hashgrid
[4]: 52da45bdb2c886f396dee93dcc521035ef5303f2