Asset pipeline related changes for Rails v5.0.3
I ran the following command using Rails v5.0.3 which is the version of
Rails for which this was last run. This time I did *not* include the
`skip-sprockets` option which seems to be the thing which determines
whether asset pipeline related changes are generated.

    rails new site --skip-bundle --skip-action-mailer \
    --skip-active-record --skip-puma --skip-action-cable \
    --skip-spring --skip-listen \
    --skip-javascript --skip-turbolinks --skip-test --force

Note that a few of the changes are not asset pipeline related and I
assume I must have just missed them first time round or perhaps they
will be removed when I do the same thing for v5.1.2 of Rails which is
the one the app is currently using.