Update generated files for Rails v5.0.3 -> v5.1.2
I upgraded Rails from v5.0.3 to v5.1.2 in this commit [1], but I
neglected to upgrade the generated files at the same time. This commit
addresses that omission. I ran the following command to generate the
changes and then accepted the ones that made sense:

    rails new site --skip-bundle --skip-action-mailer \
    --skip-active-record --skip-puma --skip-action-cable \
    --skip-spring --skip-listen \
    --skip-turbolinks --skip-test --force

Note that I intentionally left out the `skip-sprockets` and
`skip-javascript` options in order to have sensible comparisons.

[1]: 85eab323c2407f9ebb3dba9bc66f20291e41597b