Move asset-related gems into appropriate groups
* `less-rails` is needed in the `development` group so that LESS files
can be converted into a single CSS file and it is needed in the `assets`
group so that LESS files can be pre-compiled during deployment. Note
that recap `assets:precompile:default` task sets the `RAILS_GROUPS`
environment variable to `assets` when pre-compiling assets. This makes
gems in the `assets` group available for the command.

* `therubyracer` provides a JS runtime for `less-rails` and so it must
also be in both the `assets` & `development` groups.

* `uglifier` is only used to compress JavaScript as part of asset
pre-compilation on deployment and so it only needs to be in the `assets`
group.

* `yui-compressor` is only used to compress CSS as part of asset
pre-compilation on deployment and so it only needs to be in the `assets`
group.