Convert LESS -> SCSS and use sass-rails vs less-rails
I decided to do this when I discovered that using ERB in LESS files when
using `less-rails` resulted in an error. Given that we were already
using a branch of a fork of `less-rails` and it looks as if people have
been having problems getting fixes merged into `less-rails`, I decided
that this was the last straw and it was time to bite the bullet and
switch to SASS/SCSS. I achieved this as follows:

* I converted the `app/assets/stylesheets/*.less` files into their
SCSS equivalents using the `less2sass` gem [1]. It turned out that there
wasn't much difference - mainly the change in notation for global
variables.

* I then replaced the `less-rails` gem with `sass-rails` and removed the
dependency on `therubyracer` which was only there to provide
`less-rails` with a JS runtime.

Note that the removal of the dependency on `therubyracer` feels like
another benefit of making this change.

[1]: https://github.com/brauliobo/less2sass