Use sass-rails vs yui-compressor to compress CSS
Given we're already using `sass-rails`, we might as well use it as the
CSS compressor to reduce the number of dependencies.

Note that I believe the `config.assets.css_compressor` option defaults
to `:sass` as long as the `sass-rails` gem is available. However, I
elected to specify it explicitly for clarity.