Move meta tags into shared head partial
I can't see a good reason why we wouldn't want to do this on all pages
and doing this means it's included in all layouts.

The changes to the artefacts are partly due to the change in position of
the link, but also the addition of the link to pages which weren't using
the application layout.