Reject requests for unknown formats
We've been seeing a number of ActionView::MissingTemplate exceptions [1]
in production. @chrisroos was able to reproduce the exception by making
a request by setting the Accept header set to "plain/text".

This commit means that the Rails app will now respond with a 406 Not
Acceptable response as a result of the ActionController::UnknownFormat
exception that will be raised but then handled in the middleware stack.

I'm not 100% sure this is the best solution - perhaps we should respond
with HTML if we're asked for plain text, but that seems a bit odd to me.
Anyway, this change should mean that we don't see exceptions being
reported to Errbit when this happens which I think is the main thing.

[1]: https://gofreerange-errbit.herokuapp.com/apps/54e2340aed53ab1ae5000002/problems/59ef575b8fe83d000e55096f