Don't report some more BadRequest exceptions to Errbit
We were already filtering out ActionController::BadRequest exceptions
starting with "Invalid query parameters". I've just seen some similar
exceptions reported which start with "Invalid path parameters" [1].

Given that we have Errbit set to send email notifications on every
exception, it's important that we only see exceptions that require
action. I'm pretty sure we can safely ignore these exceptions and indeed
we used to
do so in the Vanilla version of the app [2].

Note that this type of exception results in an HTTP response code of 400
rather than 5XX which also suggests it's safe to ignore.

[1]:
https://gofreerange-errbit.herokuapp.com/apps/54e2340aed53ab1ae5000002/problems/5ac5b1a23fe9a300067779f6
[2]: 444182b034c5046662d9cfad454237fc8134ba93