Use GFR Video to host Assign-o-matic video
This video wasn't working for me in Google Chrome. I think this might've
been because the `movie.ogg` file was missing.

Anyway I decided to use GFR Video to transcode the video and generate
the markup to include it on gofreerange.com using video.js. Hopefully
this will mean it will now work in most browsers.