Remove CSRF meta tags from shared header
This call to `csrf_meta_tags` was originally added when the vanilla
Rails app was generated. They are only needed for generating dynamic
forms that implement non-remote links with `:method` [1]. Given that we
current don't have *any* forms, these are redundant.

This call to `csrf_meta_tags` was generating tags like this:

    <meta name="csrf-param" content="authenticity_token" />
    <meta name="csrf-token" content="Cill0vRGfvWEpLC/5TwpqXqbBlPqK7sKw5ZOXFO7DU5vBD0kznAv7k7l8fyZyw9b5mu1V5RT3d4w1PJdGAMlHw==" />

One side-effect of this was that Rails generated a different ETag for
every request to the same page, even if none of the content had changed.
Removing the tags means that Rails will generate the same ETag for a
given page thus improving the cachability of the site.

[1]:
http://api.rubyonrails.org/v5.1.6/classes/ActionView/Helpers/CsrfHelper.html#method-i-csrf_meta_tags