Ignore all exceptions handled by Rails ActionDispatch::ShowExceptions
I can't see why we'd want to report any of the exceptions which
ActionDispatch::ShowExceptions rescues and converts into suitable HTTP
response codes using ActionDispatch::ExceptionWrapper [1].

This commit changes the existing Airbrake filter so that we only report
exceptions which will not be handled by the Rails middleware.

Note that I've also opened an Airbrake issue [2] to suggest that this
should be the default behaviour for Airbrake in a Rails app.

[1]:
https://github.com/rails/rails/blob/v5.1.6/actionpack/lib/action_dispatch/middleware/exception_wrapper.rb#L9-L21
[2]: https://github.com/airbrake/airbrake/issues/834