Enable SSL for gofreerange.com
* I attempted to use the certbot tool [1] on our Linode box to do this
automatically, but I ran into a few problems and so I ended up manually
editing a number of files. It's possible this may mean we can't use the
tool to renew our certificate, but I'll investigate that separately.

* One problem was that for some reason (?) gofreerange.com is on a
different IP address than video.gofreerange.com. Previously our Apache
instance was configured (in /etc/apache2/ports.conf) to listen on port
443, but only on 178.79.133.233. Certbot didn't handle this and so I had
to manually modify this configuration to also listen on 85.159.213.6.

* Certbot duplicated the ServerAlias statement for www.gofreerange.com
into the new SSL vhost and used the %{SERVER_NAME} in the RewriteRule
target URL. However, since I didn't setup an SSL certificate for
www.gofreerange.com, this resulted in an error in the browser when
requesting http://www.gofreerange.com, because the domain did not match
the SSL certificate. I've fixed this by keep the redirect from
www.gofreerange.com -> gofreerange.com in the vhost for port 80 and
removing the ServerAlias for www.gofreerange.com in the vhost for port
443.

  * http://www.gofreerange.com -> http://gofreerange.com -> https://gofreerange.com
  * http://gofreerange.com -> https://gofreerange.com
  * https://gofreerange.com # site served successfully
  * https://www.gofreerange.com # site not served successfully (no SSL certificate)

* Certbot duplicated the whole of the
/etc/apache2/sites-available/gofreerange.com.conf vhost file to
/etc/apache2/sites-available/gofreerange.com-le-ssl.conf. This meant
there was a lot of unnecessary duplication and I found it confusing that
the SSL vhost config was separate from the port 80 vhost. So I've
combined them into the gofreerange.com.conf file and removed the
unnecessary duplication.

[1]: https://certbot.eff.org/lets-encrypt/ubuntutrusty-apache